+++ pages/cdc_costos_productos.php
@@ -255,6 +255,7 @@ desired effect
                                             <tr>
                                             <th>ID</th>
                                             <th>Nivel</th>
+                                            <th>oculto</th>
                                             <th>Nombre</th>
                                             <th style="width: 40px; text-align: center"><i class="glyphicon glyphicon-flash"></i></th>
                                             </tr>
@@ -276,6 +277,7 @@ desired effect
                                             <tr>
                                             <th>ID</th>
                                             <th>Nivel</th>
+                                            <th>oculto</th>
                                             <th>Nombre</th>
                                             <th style="width: 40px; text-align: center"><i class="glyphicon glyphicon-flash"></i></th>
                                             </tr>
@@ -297,6 +299,7 @@ desired effect
                                             <tr>
                                             <th>ID</th>
                                             <th>Nivel</th>
+                                            <th>oculto</th>
                                             <th>Nombre</th>
                                             <th class="text-center">Unidad</th>
                                             <th class="text-right">Costo unidad</th>